+++ experiments/long_runs/snowy_land.jl
@@ -59,7 +59,7 @@ outdir =
     ClimaUtilities.OutputPathGenerator.generate_output_path(diagnostics_outdir)
 
 function setup_model(FT, start_date, stop_date, Δt, domain, earth_param_set)
-    time_interpolation_method =
+    era5_time_interpolation_method =
         LONGER_RUN ? LinearInterpolation() :
         LinearInterpolation(PeriodicCalendar())
     surface_space = domain.space.surface
@@ -82,7 +82,7 @@ function setup_model(FT, start_date, stop_date, Δt, domain, earth_param_set)
         earth_param_set,
         FT;
         max_wind_speed = 25.0,
-        time_interpolation_method,
+        time_interpolation_method = era5_time_interpolation_method,
     )
 
     (; ν_ss_om, ν_ss_quartz, ν_ss_gravel) =
@@ -202,15 +202,15 @@ function setup_model(FT, start_date, stop_date, Δt, domain, earth_param_set)
     else
         modis_lai_ncdata_path = ClimaLand.Artifacts.modis_lai_multiyear_paths(;
             context = nothing,
-            start_date = start_date + Second(t0),
-            end_date = start_date + Second(t0) + Second(tf),
+            start_date,
+            end_date = stop_date,
         )
     end
     LAIfunction = ClimaLand.prescribed_lai_modis(
         modis_lai_ncdata_path,
         surface_space,
         start_date;
-        time_interpolation_method,
+        time_interpolation_method = LinearInterpolation(),
     )
     ai_parameterization =
         Canopy.PrescribedSiteAreaIndex{FT}(LAIfunction, SAI, RAI)